Replace AI slop with AI slop
ChatGPT, when I asked it to generate a zip of the repository, made up a new code instead of copying the correct one from the context window. Sure, after that there was so much context that no transformer model could just infer that they should copy that. Anyway, that's AI slop because I'm being lazy to install a C compiler or care to program it myself using a scripting language. Sure, one of the reasons I want a .exe is in case a scripting language would be blocked (which could be the case if I wrote a PowerShell script).